+++ src/gui/sparrow/main.py
@@ -414,6 +414,7 @@ class SparrowViewer(qw.QMainWindow, TalkieConnectionOwner):
                 ('Catalog', elements.CatalogState()),
                 ('Coastlines', elements.CoastlinesState()),
                 ('Borders', elements.BordersState()),
+                ('Geonames', elements.GeonamesState()),
                 ('Rivers', elements.RiversState()),
                 ('Rectangular Source', elements.SourceState()),
                 ('HUD Subtitle', elements.HudState(