+++ CHANGELOG.md
@@ -9,7 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Sparrow:
     - global colorbar handling added to CPTHandler
-
+    - added Elements: Rivers, Borders, Geonames
+- datasets:
+    - gshhg: added classes Coastlines, Rivers, Borders
+    - geonames: added Countries
 ### Fixed
 - Sparrow:
     - GeometryElement: aware to global and focal time windows for animations